\name{sddsolve}
\Rdversion{1.1}
\alias{sddsolve}
\title{
	Helper function for sdd
}
\description{
Helper function for sdd.  Not to be called directly.
}
\usage{
sddsolve(s, m)
}
\arguments{
  \item{s}{
matrix of values
}
  \item{m}{
number of rows
}
}
\details{
Helper function for sdd. Not to be called directly.
}
\value{
\item{x }{}
\item{imax }{}
\item{fmax }{}
}
\references{
}
\author{
Tamara G. Kolda, Dianne P. O'Leary
}
\note{
Not to be called directly.
}

\seealso{
}
\examples{
}